Do character replacement for credentialProfileName instead of oktaProfile
Summary:
oktaProfile variable is only non-null if you already exported the profile (manually or through bash profile), or if it's saved in ~/.okta/config.properties.
Ultimately, credentialProfileName is the variable that holds the profileName, so just do the replacement to that.

Test Plan: Remove any saved profiles and run refreshokta

Reviewers: #devops, michael.samkwakwa

Reviewed By: #devops, michael.samkwakwa

Subscribers: michael.samkwakwa

JIRA Issues: DOPS-2270

Differential Revision: https://signifyd.phacility.com/D23994